+++ .idea/misc.xml
@@ -6,11 +6,4 @@
   <component name="ProjectType">
     <option name="id" value="Android" />
   </component>
-  <component name="VisualizationToolProject">
-    <option name="state">
-      <ProjectState>
-        <option name="scale" value="0.9" />
-      </ProjectState>
-    </option>
-  </component>
 </project>
