+++ app/src/androidTest/java/com/example/sandbaks/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.homepage;
+package com.example.sandbaks;
 
 import android.content.Context;
