+++ app/src/main/java/com/example/sandbaks/DBHelper.java
@@ -1,4 +1,4 @@
-package com.example.homepage;
+package com.example.sandbaks;
 
 import android.content.ContentValues;
 import android.content.Context;
