+++ app/src/main/java/com/example/sandbaks/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.homepage;
+package com.example.sandbaks;
 
 import androidx.appcompat.app.AppCompatActivity;
