+++ app/src/main/java/com/example/sandbaks/japanFragment.java
@@ -1,4 +1,4 @@
-package com.example.homepage;
+package com.example.sandbaks;
 
 import android.os.Bundle;
 
@@ -10,10 +10,10 @@ import android.view.ViewGroup;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link stoneFragment#newInstance} factory method to
+ * Use the {@link japanFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class stoneFragment extends Fragment {
+public class japanFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +24,7 @@ public class stoneFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public stoneFragment() {
+    public japanFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +34,11 @@ public class stoneFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment stoneFragment.
+     * @return A new instance of fragment japanFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static stoneFragment newInstance(String param1, String param2) {
-        stoneFragment fragment = new stoneFragment();
+    public static japanFragment newInstance(String param1, String param2) {
+        japanFragment fragment = new japanFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +59,6 @@ public class stoneFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_stone, container, false);
+        return inflater.inflate(R.layout.fragment_japan, container, false);
     }
 }
