+++ app/src/main/java/com/example/sandbaks/stoneFragment.java
@@ -1,4 +1,4 @@
-package com.example.homepage;
+package com.example.sandbaks;
 
 import android.os.Bundle;
 
@@ -10,10 +10,10 @@ import android.view.ViewGroup;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link bronzeFragment#newInstance} factory method to
+ * Use the {@link stoneFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class bronzeFragment extends Fragment {
+public class stoneFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +24,7 @@ public class bronzeFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public bronzeFragment() {
+    public stoneFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +34,11 @@ public class bronzeFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment bronzeFragment.
+     * @return A new instance of fragment stoneFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static bronzeFragment newInstance(String param1, String param2) {
-        bronzeFragment fragment = new bronzeFragment();
+    public static stoneFragment newInstance(String param1, String param2) {
+        stoneFragment fragment = new stoneFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +59,6 @@ public class bronzeFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_bronze, container, false);
+        return inflater.inflate(R.layout.fragment_stone, container, false);
     }
 }
