+++ app/src/main/res/layout/activity_playgame.xml
@@ -151,7 +151,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragmentContainerView"
-            android:name="com.example.homepage.stoneFragment"
+            android:name="com.example.sandbaks.stoneFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:visibility="visible"
