+++ app/src/test/java/com/example/sandbaks/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.homepage;
+package com.example.sandbaks;
 
 import org.junit.Test;
